int fix

+++ plugin.py
@@ -837,7 +837,7 @@ def set_scale(device_functions, actual_function_name, raw):
             if item['code'] == actual_function_name:
                 the_values = json.loads(item['values'])
                 scale = int(the_values.get('scale', 0))
-    return raw * 10 if scale == 1 else raw
+    return int(raw * 10) if scale == 1 else raw
 
 def get_scale(device_functions, actual_function_name, raw):
     scale = 0
@@ -848,7 +848,7 @@ def get_scale(device_functions, actual_function_name, raw):
             if item['code'] == actual_function_name:
                 the_values = json.loads(item['values'])
                 scale = the_values.get('scale', 0)
-    return raw / 10 if scale == 1 else raw
+    return float(raw / 10) if scale == 1 else raw
 
 def rgb_to_hsv(r, g, b):
     h, s, v = colorsys.rgb_to_hsv(r / 255, g / 255, b / 255)
